+++ src/backend/core/api/feature_flag.py
@@ -15,6 +15,7 @@ class FeatureFlag:
         "subtitle": "ROOM_SUBTITLE_ENABLED",
         "file_upload": "FILE_UPLOAD_ENABLED",
         "addons": "ADDONS_ENABLED",
+        "application": "APPLICATION_ENABLED",
     }
 
     @classmethod